moved fixXreacs.py from python to python/moose for packaging, changes made to get fixXreacs files
